requirements.txt
torch
diffusers
transformers
accelerate
safetensors
matplotlib
pillow
huggingface_hub
datasets
opencv-python
rembg